FIRST PUBLISHING OF PORTFOLIO!!!!! WOOOOOO!!!!!
Keep in mind. I have not yet figured out how to do the form. That shouldn't be terribly hard, but I'm ready to publish the site, so I will figure that out in a minute. Also, I would like to copy the entire index and css file by hand just to clean the code.